Add TextWithIconCell
